Update file layout; separate agent/device
