+++ datto/base.py
@@ -1,19 +1,33 @@
-class DattoAsset():
-    '''Class to normalize a Datto Asset 
-    as an object'''
-    def __init__(self, agent):
-        self.name = agent['name']
-        self.local_ip = agent['localIp']
-        self.os = agent['os']
-        self.unprotected_volumes = agent['unprotectedVolumeNames']
-        self.agent_version = agent['agentVersion']
-        self.is_paused = agent['isPaused']
-        self.is_archived = agent['isArchived']
-        self.latest_offsite = agent['latestOffsite']
-        self.last_snapshot = agent['lastSnapshot']
-        self.last_screenshot_attempt = agent['lastScreenshotAttempt']
-        self.last_screenshot_attempt_status = agent['lastScreenshotAttemptStatus']
-        self.last_screenshot_url = agent['lastScreenshotUrl']
-        self.fqdn = agent['fqdn']
-        self.backups = agent['backups']
-        self.type = agent['type']
+
+class Base():
+    """Base methods for various classes"""
+
+    def __init__(self):
+        pass
+
+    def display_time(self, seconds, granularity=2):
+        """
+        Converts an integer (number of seconds) into a readable time format with certain granularity.
+
+        From "Mr. B":
+        https://stackoverflow.com/questions/4048651/python-function-to-convert-seconds-into-minutes-hours-and-days/24542445#answer-24542445
+        """
+        intervals = (
+            ('weeks', 604800),  # 60 * 60 * 24 * 7
+            ('days', 86400),    # 60 * 60 * 24
+            ('hours', 3600),    # 60 * 60
+            ('minutes', 60),
+            ('seconds', 1),
+        )
+
+        seconds = int(seconds)
+        result = []
+
+        for name, count in intervals:
+            value = seconds // count
+            if value:
+                seconds -= value * count
+                if value == 1:
+                    name = name.rstrip('s')
+                result.append("{} {}".format(value, name))
+        return ', '.join(result[:granularity])
