+++ main.py
@@ -2,20 +2,17 @@
 
 # Import
 import sys
-import argparse
+from argparse import ArgumentParser
 import logging
-from datto import DattoCheck
+import datto
 import config
 
 from logging import StreamHandler, DEBUG, INFO, Formatter
 from logging.handlers import RotatingFileHandler
 
 def main():
-    """Main entry.
+    """Main"""
 
-    Parse command line arguments.
-    Basic checks on CLI input.
-    Initialize and run."""
     __authors__ = ['Tommy Harris', 'Ryan Shoemaker']
     __date__ = 'September 8, 2019'
     __description__ = """Using the Datto API, get information on current status \
@@ -23,30 +20,24 @@ def main():
 
     To send the results as an email, provide the optional email parameters."""
 
-    parser = argparse.ArgumentParser(description=__description__,
-                                     epilog='Developed by {} on \
-                                     {}'.format(", ".join(__authors__), __date__))
+    parser = ArgumentParser(description=__description__,
+                            epilog='Developed by {} on \
+                            {}'.format(", ".join(__authors__), __date__))
 
     # "Optional" arguments
     parser.add_argument('-v', '--verbose',
                         help='Print verbose output to stdout',
                         action='store_true')
-    parser.add_argument('--unprotected-volumes', '-u', help='Include \
+    parser.add_argument('-u', '--unprotected-volumes', help='Include \
         any unprotected volumes in the final report',
         action='store_true')
 
     args = parser.parse_args()
 
-    # args sanity check
-    if args.send_email:
-        if not args.email_from or not args.email_to or not args.mx_endpoint:
-            raise InvalidEmailSettings("You must have at least a \
-            sender, recipient, and MX endpoint.")
-
     # Add rotating log
     logger = logging.getLogger("Datto Check")
     logger.setLevel(DEBUG)
-    handler = RotatingFileHandler(LOG_FILE, maxBytes=30000, backupCount=3)
+    handler = RotatingFileHandler(config.LOG_FILE, maxBytes=30000, backupCount=3)
     handler.setLevel(INFO)
     formatter = Formatter('%(asctime)s - %(levelname)s - %(message)s')
     handler.setFormatter(formatter)
@@ -60,7 +51,7 @@ def main():
         handler.setFormatter(formatter)
         logger.addHandler(handler)
 
-    datto_check = DattoCheck(args)
+    datto_check = datto.DattoCheck(args)
     datto_check.run()
     return 0
